+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Favorites: Fixed gear icon to manage the favorites.
+- Core: Multiselect grid, Added check if the tools property is given.
 - Tasks/Start page: Put headers back so the startpage widget can be sorted on date.
 - Calendar: Fixed problem that opening event dialog the 2nd time did not work.
 - Core: Don't search in TEXT fields by default